multiple_pdf_chabot/requirements.txt
langchain==0.3.26
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.47.0
openai==1.97.0
faiss-cpu==1.9.0
altair==5.5.0
tiktoken==0.4.0
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers== 5.0.0